fix: merged skill_analyzer and SkillAnalyzer
