+++ src/analyzers/ProjectAnalyzer.py
@@ -16,7 +16,7 @@ from src.analyzers.ContributionAnalyzer import ContributionAnalyzer, Contributio
 from utils.RepoFinder import RepoFinder
 from src.ProjectManager import ProjectManager
 from src.Project import Project
-from src.analyzers.skill_analyzer import SkillAnalyzer
+from src.analyzers.SkillAnalyzer import SkillAnalyzer
 from src.analyzers.code_metrics_analyzer import CodeMetricsAnalyzer
 from src.generators.ResumeInsightsGenerator import ResumeInsightsGenerator
 from src.ConfigManager import ConfigManager
