+++ src/analyzers/SkillAnalyzer.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Any, Dict, Iterable, List, Callable, Tuple
+from typing import Any, Dict, Iterable, List, Callable, Tuple, Set
 
 import os
 
@@ -12,10 +12,35 @@ from .skill_patterns import DEP_TO_SKILL, SNIPPET_PATTERNS, KNOWN_CONFIG_HINTS
 from .skill_proficiency import ProficiencyEstimator
 from .code_metrics_analyzer import CodeMetricsAnalyzer, CodeFileAnalysis
 
+# Heuristic mapping: which snippet-based skills make sense for which languages.
+# This lets us avoid running JS regexes on Python files, etc.
+LANG_TO_ALLOWED_SNIPPET_SKILLS: Dict[str, Set[str]] = {
+    "python": {"Django", "Flask", "FastAPI", "Python"},
+    "javascript": {"React", "Next.js"},
+    "typescript": {"React", "Next.js"},
+    "js": {"React", "Next.js"},
+    "ts": {"React", "Next.js"},
+    "c++": {"C++", "CMake"},
+    "cpp": {"C++", "CMake"},
+    "c": {"CMake"},
+    "c#": {"C#", "CMake"},
+    "java": {"Java", "CMake"},
+    "rust": {"Rust", "CMake"},
+}
+
 
 class SkillAnalyzer:
     """
     High-level skill analysis for a single extracted project directory.
+
+    Responsibilities:
+    - Run CodeMetricsAnalyzer to compute per-file metrics
+    - Extract Evidence objects from:
+      - languages used (via file extensions / language detection),
+      - dependencies (requirements.txt, package.json, pom.xml, etc.),
+      - config file conventions (Dockerfile, next.config.js, etc.),
+      - snippet patterns in code (import lines, includes, etc.).
+    - Aggregate Evidence into SkillProfileItem objects with proficiency scores.
     """
 
     def __init__(self, root_dir: Path) -> None:
@@ -139,6 +164,17 @@ class SkillAnalyzer:
     # ------------------------------------------------------------------
 
     def analyze(self) -> Dict[str, Any]:
+        """
+        Perform full skill analysis.
+
+        Returns:
+            {
+              "skills": List[SkillProfileItem],
+              "stats": Dict[str, Any],   # project-level metrics used for scoring
+              "dimensions": Dict[str, Any],
+              "resume_suggestions": List[Dict[str, Any]],
+            }
+        """
         file_analyses = self.metrics_analyzer.analyze()
         stats = self._build_stats(file_analyses)
 
@@ -153,7 +189,7 @@ class SkillAnalyzer:
         skills = self._build_skill_profiles(evidence, stats)
         dimensions = self._compute_dimensions(stats)
 
-        # Placeholder for future resume/portfolio bullets built directly here
+        # placeholder for future fully-formed bullets
         resume_suggestions: List[Dict[str, Any]] = []
 
         return {
@@ -164,28 +200,56 @@ class SkillAnalyzer:
         }
 
     # ------------------------------------------------------------------
-    # Stats used by ProficiencyEstimator
+    # Stats used by ProficiencyEstimator / dimensions
     # ------------------------------------------------------------------
 
     def _build_stats(self, file_analyses: List[CodeFileAnalysis]) -> Dict[str, Any]:
+        """
+        Produce a stats dict from code metrics.
+        """
         summary = self.metrics_analyzer.summarize(file_analyses)
         overall = summary.get("overall", {})
         per_lang = summary.get("per_language", {})
 
-        return {
+        stats: Dict[str, Any] = {
             "overall": overall,
             "per_language": per_lang,
         }
 
+        # Provide a Python-specific section if Python files exist
+        py_key = None
+        for lang_name in per_lang.keys():
+            if lang_name and lang_name.lower().startswith("python"):
+                py_key = lang_name
+                break
+
+        if py_key:
+            py_data = per_lang[py_key]
+            stats["python"] = {
+                "files": py_data.get("file_count", 0),
+                "loc": py_data.get("loc", 0),
+                "functions": py_data.get("functions", 0),
+                "comment_ratio": py_data.get("comment_ratio", 0.0),
+                "avg_functions_per_file": py_data.get("avg_functions_per_file", 0.0),
+            }
+
+        # Could add similar language-specific blocks for JS/TS, C++, etc. later
+        return stats
+
     # ------------------------------------------------------------------
-    # Evidence extractors
+    # Evidence extraction
     # ------------------------------------------------------------------
 
     def _language_evidence(
-        self, file_analyses: List[CodeFileAnalysis]
+        self, file_analyses: Iterable[CodeFileAnalysis]
     ) -> List[Evidence]:
+        """
+        Evidence derived from file extensions / detected languages.
+        Uses LOC-weighted scoring for better proficiency estimates.
+        """
         evidence: List[Evidence] = []
 
+        # First pass: gather LOC per language
         loc_per_lang: Dict[str, int] = {}
         for fa in file_analyses:
             lang = fa.language
@@ -198,7 +262,7 @@ class SkillAnalyzer:
 
         max_loc = max(loc_per_lang.values()) or 1
 
-        # TAXONOMY might be a dict or a set; tests clearly have it as a set.
+        # TAXONOMY might be a dict or a set; handle both cases
         is_tax_dict = isinstance(TAXONOMY, dict)
 
         for lang, loc in loc_per_lang.items():
@@ -227,6 +291,9 @@ class SkillAnalyzer:
         return evidence
 
     def _dependency_evidence(self) -> List[Evidence]:
+        """
+        Walks text-like dependency/config files and applies DEP_TO_SKILL patterns.
+        """
         evidence: List[Evidence] = []
         text_like_exts = {
             ".txt",
@@ -271,6 +338,9 @@ class SkillAnalyzer:
         return evidence
 
     def _config_evidence(self) -> List[Evidence]:
+        """
+        Evidence from config-file naming conventions (e.g., next.config.js, Dockerfile).
+        """
         evidence: List[Evidence] = []
 
         for path in self._iter_project_files():
@@ -294,34 +364,75 @@ class SkillAnalyzer:
         return evidence
 
     def _extract_snippet_skills(self, file_analyses: List[CodeFileAnalysis]) -> None:
+        """
+        For each code file, scan its text once and record which snippet patterns matched.
+        This avoids re-reading files in _snippet_evidence and also gates patterns
+        by language to keep Analyze Skills fast.
+        
+        Uses language gating from LANG_TO_ALLOWED_SNIPPET_SKILLS to avoid running
+        JS regexes on Python files, etc.
+        """
         for fa in file_analyses:
+            full_path = fa.path
+            
+            if not isinstance(full_path, Path):
+                full_path = Path(full_path)
             try:
-                text = fa.path.read_text(encoding="utf-8", errors="ignore")
+                text = full_path.read_text(encoding="utf-8", errors="ignore")
             except OSError:
                 continue
+            lang = (fa.language or "").lower()
+            allowed_skills = LANG_TO_ALLOWED_SNIPPET_SKILLS.get(lang, [])
 
-            for pattern, skill in self._iter_pattern_pairs(SNIPPET_PATTERNS):
-                if hasattr(pattern, "findall"):
-                    matches = len(pattern.findall(text))
-                else:
-                    matches = text.count(str(pattern))
-                if matches > 0:
-                    fa.snippet_matches[skill] = (
-                        fa.snippet_matches.get(skill, 0) + matches
-                    )
+            if not allowed_skills:
+                continue
 
-    def _snippet_evidence(self, file_analyses: List[CodeFileAnalysis]) -> List[Evidence]:
+            for skill in allowed_skills:
+            # Count matches per skill
+                for pattern, skill in self._iter_pattern_pairs(SNIPPET_PATTERNS):
+                    # If we have a whitelist for this language, skip skills that don't apply
+                    if allowed_skills is not None and skill not in allowed_skills:
+                        continue
+                    
+                    if hasattr(pattern, "findall"):
+                        matches = len(pattern.findall(text))
+                    else:
+                        matches = text.count(str(pattern))
+                    
+                    if matches > 0:
+                        fa.snippet_matches[skill] = (
+                            fa.snippet_matches.get(skill, 0) + matches
+                        )
+
+            # Also populate snippet_skills list for backward compatibility
+            if fa.snippet_matches:
+                fa.snippet_skills.extend(sorted(fa.snippet_matches.keys()))
+
+    def _snippet_evidence(
+        self, file_analyses: Iterable[CodeFileAnalysis]
+    ) -> List[Evidence]:
+        """
+        Turn snippet_matches on each file into Evidence entries.
+        Weight increases with match count.
+        """
         evidence: List[Evidence] = []
-
+        
         for fa in file_analyses:
+            if not fa.snippet_matches:
+                continue
+
             for skill, count in fa.snippet_matches.items():
                 weight = min(1.0, 0.3 + 0.1 * count)
+                try:
+                    file_path = str(fa.path.resolve().relative_to(self.root_dir.resolve()))
+                except ValueError:
+                    file_path = str(fa.path.resolve())
                 evidence.append(
                     Evidence(
                         skill=skill,
                         source="snippet_pattern",
                         raw=f"{skill} x{count}",
-                        file_path=str(fa.path.relative_to(self.root_dir)),
+                        file_path=file_path,
                         weight=weight,
                     )
                 )
@@ -329,91 +440,161 @@ class SkillAnalyzer:
         return evidence
 
     # ------------------------------------------------------------------
-    # Aggregation into SkillProfileItems
+    # Building SkillProfileItem objects
     # ------------------------------------------------------------------
 
     def _build_skill_profiles(
-        self, evidence_list: List[Evidence], stats: Dict[str, Any]
+        self, evidence: List[Evidence], stats: Dict[str, Any]
     ) -> List[SkillProfileItem]:
+        """
+        Aggregate evidence into SkillProfileItem objects with proficiency estimates.
+        """
         by_skill: Dict[str, List[Evidence]] = {}
-        for ev in evidence_list:
-            by_skill.setdefault(ev.skill, []).append(ev)
+        for e in evidence:
+            by_skill.setdefault(e.skill, []).append(e)
 
         profiles: List[SkillProfileItem] = []
-
-        for skill, evs in by_skill.items():
-            proficiency = self.prof_estimator.estimate(skill, evs, stats)
-            confidence = round(min(1.0, 0.3 + 0.1 * len(evs) + 0.1 * proficiency), 2)
-
+        for skill, ev_list in by_skill.items():
+            if not skill:
+                continue
+            
+            # Estimate proficiency using ProficiencyEstimator
+            proficiency = self.prof_estimator.estimate(skill, ev_list, stats)
+            confidence = round(min(1.0, 0.3 + 0.1 * len(ev_list) + 0.1 * proficiency), 2)
+            
             profiles.append(
                 SkillProfileItem(
                     skill=skill,
                     proficiency=proficiency,
                     confidence=confidence,
-                    evidence=evs,
+                    evidence=ev_list,
                 )
             )
 
-        profiles.sort(key=lambda s: (s.proficiency, s.confidence), reverse=True)
+        # Sort by proficiency, then confidence
+        profiles.sort(
+            key=lambda p: (p.proficiency, p.confidence),
+            reverse=True,
+        )
         return profiles
 
     # ------------------------------------------------------------------
-    # Dimensions (meta-skills)
+    # Internal helpers: dimensions
     # ------------------------------------------------------------------
 
-    def _compute_dimensions(self, stats: Dict[str, Any]) -> Dict[str, Dict[str, Any]]:
+    def _compute_dimensions(self, stats: Dict[str, Any]) -> Dict[str, Any]:
+        """
+        Compute meta-skill dimensions from project statistics.
+        """
         overall = stats.get("overall", {}) or {}
         per_lang = stats.get("per_language", {}) or {}
 
-        # Testing discipline: ratio of test files to total files
+        # --- Testing discipline ---
         total_files = overall.get("total_files", overall.get("file_count", 0)) or 1
-
         total_test_files = 0
+        
         if isinstance(per_lang, dict):
             for lang_stats in per_lang.values():
                 if isinstance(lang_stats, dict):
                     total_test_files += int(lang_stats.get("test_file_count", 0))
+        
+        # Also check for test_file_ratio in overall if available (from skill_analyzer.py)
+        test_ratio = overall.get("test_file_ratio")
+        if test_ratio is None:
+            test_ratio = total_test_files / total_files if total_files > 0 else 0.0
+        
+        test_score = min(1.0, test_ratio / 0.4)  # 0.4+ tests/code ~= strong
+        testing_level = self._level_from_score(test_score)
 
-        testing_score = total_test_files / total_files
-
-        # Documentation habits: use comment_ratio from overall if available
-        doc_score = float(overall.get("comment_ratio", 0.5))
-
-        # Modularity: inverse of max function length, clamped
-        max_func = overall.get("max_function_length", 0)
-        modularity_score = 1.0 / (1.0 + max_func / 50.0) if max_func else 0.5
-
-        # Language depth: number of languages with significant LOC
-        lang_count = len(per_lang) if isinstance(per_lang, dict) else 0
-        lang_depth_score = min(1.0, lang_count / 4.0)
-
-        return {
-            "testing_discipline": {
-                "score": testing_score,
-                "level": self._level_from_score(testing_score),
-                "raw": {
-                    "total_files": total_files,
-                    "total_test_files": total_test_files,
-                },
-            },
-            "documentation_habits": {
-                "score": doc_score,
-                "level": self._level_from_score(doc_score),
-                "raw": {},
-            },
-            "modularity": {
-                "score": modularity_score,
-                "level": self._level_from_score(modularity_score),
-                "raw": {"max_function_length": max_func},
-            },
-            "language_depth": {
-                "score": lang_depth_score,
-                "level": self._level_from_score(lang_depth_score),
-                "raw": {"language_count": lang_count},
+        testing_dim = {
+            "score": round(test_score, 2),
+            "level": testing_level,
+            "raw": {
+                "test_file_ratio": test_ratio,
+                "num_test_files": total_test_files,
+                "num_code_files": overall.get("num_code_files", total_files),
+                "total_files": total_files,
+                "total_test_files": total_test_files,
             },
         }
 
+        # --- Documentation habits ---
+        comment_ratio = overall.get("comment_ratio", 0.0)
+        # e.g. 0.05 is minimal, 0.2+ is solid
+        doc_score = min(1.0, comment_ratio / 0.2)
+        doc_level = self._level_from_score(doc_score)
+
+        doc_dim = {
+            "score": round(doc_score, 2),
+            "level": doc_level,
+            "raw": {
+                "comment_ratio": comment_ratio,
+            },
+        }
+
+        # --- Modularity ---
+        avg_funcs = overall.get("avg_functions_per_file", 0.0)
+        max_func_len = overall.get("max_function_length", 0)
+
+        # heuristic: more functions and shorter max function is better
+        modularity_score = 0.0
+        if avg_funcs >= 3:
+            modularity_score += 0.5
+        if max_func_len <= 50:
+            modularity_score += 0.5
+        elif max_func_len <= 100:
+            modularity_score += 0.25
+
+        modularity_score = min(1.0, modularity_score)
+        modularity_level = self._level_from_score(modularity_score)
+
+        modularity_dim = {
+            "score": round(modularity_score, 2),
+            "level": modularity_level,
+            "raw": {
+                "avg_functions_per_file": avg_funcs,
+                "max_function_length": max_func_len,
+            },
+        }
+
+        # --- Language depth ---
+        total_loc = sum(lang_stats.get("loc", 0) for lang_stats in per_lang.values() if isinstance(lang_stats, dict))
+        # A simple proxy: fraction of languages above a LOC threshold
+        depth_languages = {
+            lang: data["loc"]
+            for lang, data in per_lang.items()
+            if isinstance(data, dict) and data.get("loc", 0) >= 500  # arbitrary "non-toy" threshold
+        }
+        
+        lang_count = len(per_lang) if isinstance(per_lang, dict) else 0
+        depth_ratio = len(depth_languages) / max(1, lang_count) if lang_count > 0 else 0.0
+        
+        lang_depth_score = min(
+            1.0, 0.5 + depth_ratio * 0.5
+        )  # base 0.5 for having code at all
+        lang_depth_level = self._level_from_score(lang_depth_score)
+
+        lang_depth_dim = {
+            "score": round(lang_depth_score, 2),
+            "level": lang_depth_level,
+            "raw": {
+                "languages": {
+                    lang: {"loc": loc} for lang, loc in depth_languages.items()
+                },
+                "total_loc": total_loc,
+                "language_count": lang_count,
+            },
+        }
+
+        return {
+            "testing_discipline": testing_dim,
+            "documentation_habits": doc_dim,
+            "modularity": modularity_dim,
+            "language_depth": lang_depth_dim,
+        }
+
     def _level_from_score(self, score: float) -> str:
+        """Convert numeric score to human-readable level."""
         if score >= 0.75:
             return "strong"
         if score >= 0.5:
