+++ tests/test_skill_analyzer.py
@@ -9,8 +9,8 @@ SRC = ROOT / "src"
 if str(SRC) not in sys.path:
     sys.path.insert(0, str(SRC))
 
-from analyzers.skill_analyzer import SkillAnalyzer  # type: ignore
-from analyzers.skill_models import SkillProfileItem  # type: ignore
+from src.analyzers.SkillAnalyzer import SkillAnalyzer
+from src.analyzers.skill_models import SkillProfileItem
 
 
 def _write_file(path: Path, contents: str) -> None:
